+++ hack/image-updates.sh
@@ -4,6 +4,8 @@ YQ="${YQ:-yq}"
 KUSTOMIZE="${KUSTOMIZE:-kustomize}"
 CRANE="${CRANE:-crane}"
 
+NEED_CHART_UPDATE=n
+
 for CONFIG_FILE in ./config/manager/*_images.yaml; do
 	BASE="$(yq -e .base "${CONFIG_FILE}")"
 	for KEY in $(yq -e ".components | keys | .[]" "${CONFIG_FILE}"); do
@@ -12,8 +14,13 @@ for CONFIG_FILE in ./config/manager/*_images.yaml; do
 		if [ "$(yq -e ".components.${KEY}.tag" "${CONFIG_FILE}" )" != "${TAG}" ] ; then
 			echo "⬆ ${IMG}"
 			yq -ie ".components.${KEY}.tag = \"${TAG}\"" "${CONFIG_FILE}"
+			NEED_CHART_UPDATE=y
 		else
 			echo "= ${IMG}"
 		fi
 	done
 done
+
+if [ "$NEED_CHART_UPDATE" = y ]; then
+	make sync-chart
+fi