+++ CHANGELOG.md
@@ -17,8 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed service resources relying on default protocol version.
 - Moved NetworkPolicy for DRBD out of default deployed resources.
 - Updated images:
-  * LINSTOR 1.25.0
-  * DRBD Reactor 1.3.0
+  * LINSTOR 1.25.1
+  * LINSTOR CSI 1.3.0
+  * DRBD Reactor 1.4.0
   * Latest CSI sidecars
 - Add a default toleration for the HA Controller taints to the operator.